requirements.txt
streamlit
aif360
shap
plotly
pandas
scikit-learn
matplotlib
seaborn
fpdf2
openpyxl
numpy
requests
